allowedContracts v1
In this updated codebase, the DAO contract has been enhanced with the ability to dictate which contracts the issued NFTs are allowed to interact with. The addAllowedContract function allows authorized users to add a contract address to the allowedContracts mapping, and the isAllowed function checks if a given contract is in the mapping. The addAllowedContract function can only be called after a successful proposal vote, as it requires that there is a proposal with the most votes and that the proposal has the majority of the votes.

This new functionality allows the DAO to control which contracts the NFTs are allowed to interact with, for example, OpenSea or NFT fractionalization contracts. This can be useful for ensuring that the NFTs are only used for legitimate purposes and to prevent them from being used in any malicious or unauthorized ways.